chore(story): move Parser.clean_source to Story.clean_source
