+++ storyscript/parser/parser.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
-import re
-
 from lark import Lark
 from lark.exceptions import UnexpectedInput, UnexpectedToken
 
@@ -18,12 +16,6 @@ class Parser:
     def __init__(self, algo='lalr', ebnf_file=None):
         self.algo = algo
         self.ebnf_file = ebnf_file
-
-    @staticmethod
-    def clean_source(source):
-        expression = '(?<=###)\s(.*|\\n)+(?=\s###)|#(.*)'
-        source = re.sub(expression, '', source)
-        return '{}\n'.format(source)
 
     def indenter(self):
         """
@@ -53,7 +45,6 @@ class Parser:
         """
         Parses the source string.
         """
-        source = self.clean_source(source)
         lark = self.lark()
         try:
             tree = lark.parse(source)
