+++ tests/unittests/parser/parser.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 import os
-import re
 
 from lark import Lark
 from lark.exceptions import UnexpectedInput, UnexpectedToken
@@ -24,14 +23,6 @@ def ebnf_file(request):
     def teardown():
         os.remove('test.ebnf')
     request.addfinalizer(teardown)
-
-
-def test_parser_clean_source(patch):
-    patch.object(re, 'sub')
-    result = Parser.clean_source('source')
-    expression = '(?<=###)\s(.*|\\n)+(?=\s###)|#(.*)'
-    re.sub.assert_called_with(expression, '', 'source')
-    assert result == '{}\n'.format(re.sub())
 
 
 def test_parser_init(parser):
@@ -86,10 +77,9 @@ def test_parser_parse(patch, parser):
     """
     Ensures the build method can build the grammar
     """
-    patch.many(Parser, ['clean_source', 'lark', 'transformer'])
+    patch.many(Parser, ['lark', 'transformer'])
     result = parser.parse('source')
-    Parser.clean_source.assert_called_with('source')
-    Parser.lark().parse.assert_called_with(Parser.clean_source())
+    Parser.lark().parse.assert_called_with('source')
     Parser.transformer().transform.assert_called_with(Parser.lark().parse())
     assert result == Parser.transformer().transform()
 
@@ -97,7 +87,7 @@ def test_parser_parse(patch, parser):
 def test_parser_parser_unexpected_token(capsys, patch, magic, parser):
     patch.init(StoryError)
     patch.object(StoryError, 'message')
-    patch.many(Parser, ['clean_source', 'lark', 'transformer'])
+    patch.many(Parser, ['lark', 'transformer'])
     Parser.lark().parse.side_effect = UnexpectedToken(magic(), 'exp', 0, 1)
     with raises(SystemExit):
         parser.parse('source', debug=False)
@@ -106,7 +96,7 @@ def test_parser_parser_unexpected_token(capsys, patch, magic, parser):
 
 
 def test_parser_parser_unexpected_token_debug(patch, magic, parser):
-    patch.many(Parser, ['clean_source', 'lark', 'transformer'])
+    patch.many(Parser, ['lark', 'transformer'])
     Parser.lark().parse.side_effect = UnexpectedToken(magic(), 'exp', 0, 1)
     with raises(UnexpectedToken):
         parser.parse('source', debug=True)
@@ -115,7 +105,7 @@ def test_parser_parser_unexpected_token_debug(patch, magic, parser):
 def test_parser_parser_unexpected_input(capsys, patch, magic, parser):
     patch.init(StoryError)
     patch.object(StoryError, 'message')
-    patch.many(Parser, ['clean_source', 'lark', 'transformer'])
+    patch.many(Parser, ['lark', 'transformer'])
     Parser.lark().parse.side_effect = UnexpectedInput(magic(), 0, 0, 0)
     with raises(SystemExit):
         parser.parse('source', debug=False)
@@ -124,7 +114,7 @@ def test_parser_parser_unexpected_input(capsys, patch, magic, parser):
 
 
 def test_parser_parser_unexpected_input_debug(patch, magic, parser):
-    patch.many(Parser, ['clean_source', 'lark', 'transformer'])
+    patch.many(Parser, ['lark', 'transformer'])
     Parser.lark().parse.side_effect = UnexpectedInput(magic(), 0, 0, 0)
     with raises(UnexpectedInput):
         parser.parse('source', debug=True)
